+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `period` will be documented in this file
 
+## 1.4.5 - 2020-02-05
+
+- Fix for PeriodCollection::gaps() with excluded boundaries (#58)
+
 ## 1.4.4 - 2019-08-05
 
 - ~Performance improvement in `Period::contains()` (#46)~ edit: this change wasn't merged and targeted at 2.0